requirements.txt
APScheduler==3.11.2
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
psycopg2-binary==2.9.11
pyTelegramBotAPI==4.31.0
requests==2.32.5
tzlocal==5.3.1
urllib3==2.6.3

python-dotenv